+++ BoardConfigs.cmake
@@ -341,7 +341,6 @@ elseif ( HW_CONFIG EQUAL 8 )
         set(GPIOHSTXD2 19 CACHE STRING "HSTX D2+ pin")
         set(GPIOHSTXCK 13 CACHE STRING "HSTX CK+ pin")
         set(GPIOHSTXINVERTED 1 CACHE STRING "Set to 1 if HSTX pins are inverted: D- = D+ - 1, not inverted (default): D- = D+ + 1")
-        set(USEPICOHDMI 1 CACHE BOOL "Set to 1 to use PICO HDMI with HSTX")
    endif()
    set(ENABLE_VU_METER 1 CACHE BOOL "Enable VU meter (fruitjam NeoPixel leds)")
    # Override default UART settings used for printf debugging on the Adafruit Fruit Jam board.
@@ -527,7 +526,6 @@ elseif ( HW_CONFIG EQUAL 13 )
         set(GPIOHSTXD1 17 CACHE STRING "HSTX D1+ pin")  
         set(GPIOHSTXD2 19 CACHE STRING "HSTX D2+ pin")
         set(GPIOHSTXINVERTED 1 CACHE STRING "Set to 1 if HSTX pins are inverted: D- = D+ - 1, not inverted (default): D- = D+ + 1")
-        set(USEPICOHDMI 1 CACHE BOOL "Set to 1 to use PICO HDMI with HSTX")
    endif()
    set(PSRAM_CS_PIN 8 CACHE STRING "Select the GPIO pin for PSRAM chip select")
 elseif ( HW_CONFIG EQUAL 14 )
@@ -595,9 +593,7 @@ if (NOT DEFINED GPIOHSTXD0)
     set(GPIOHSTXCK 0 CACHE STRING "HSTX CK pin")
     set(GPIOHSTXINVERTED 0 CACHE STRING "Set to 1 if HSTX pins are inverted: D- = D+ - 1, not inverted (default): D- = D+ + 1")
 endif()
-if (NOT DEFINED USEPICOHDMI)
-    set(USEPICOHDMI 0 CACHE BOOL "Set to 1 to use PICO HDMI with HSTX")
-endif()
+
 if (NOT DEFINED ENABLE_PIO_USB ) 
     set(ENABLE_PIO_USB 0 CACHE BOOL "Enable PIO USB support")
 endif()
@@ -684,7 +680,6 @@ if ( NOT GPIOHSTXD0 EQUAL 0 )
     message("HSTX D2+ pin           : ${GPIOHSTXD2}")
     message("HSTX CK+ pin           : ${GPIOHSTXCK}")
     message("HSTX Inverted         : ${GPIOHSTXINVERTED}")
-    message("Use PICO HDMI         : ${USEPICOHDMI}")
 else()
     if ( FORCE_DVI EQUAL 0)
         message("HSTX is not enabled, using default display output.")
