+++ CMakeLists.txt
@@ -24,13 +24,16 @@ FlashParams.cpp
 add_subdirectory(drivers/pico_fatfs)     
 add_subdirectory(drivers/pico_audio_i2s)
 add_subdirectory(drivers/lwmem)
-add_subdirectory(drivers/pico_hstx)
+# add_subdirectory(drivers/pico_hstx)
 # add_subdirectory(drivers/pico_audio_mcp4822) # SPI Audio using MCP4822 DAC. Works but not used
 add_subdirectory(drivers/pico_psram)
 add_subdirectory(drivers/pico_hdmi)
 
 target_include_directories(pico_shared INTERFACE
         .
+)
+target_compile_definitions(pico_shared INTERFACE 
+ NOHSTXCLOCK=1 # Do not set the hstx clock in the pico_hdmi driver, as it is already set FrensHelpers.cpp
 )
 pico_generate_pio_header(${projectname} ${CMAKE_CURRENT_LIST_DIR}/ws2812.pio)
 target_link_libraries(pico_shared 
@@ -39,7 +42,7 @@ target_link_libraries(pico_shared
                       pico_fatfs 
                       pico_audio_i2s 
                       lwmem 
-                      pico_hstx 
+                      #pico_hstx 
                       #pico_audio_mcp4822 
                       pico_psram
                       pico_hdmi)
