+++ drivers/pico_hdmi/hstx_data_island_queue.h
@@ -5,9 +5,7 @@
 
 #include <stdbool.h>
 #include <stdint.h>
-#ifdef __cplusplus
-extern "C" {
-#endif
+
 /**
  * Initialize the Data Island queue and scheduler.
  */
@@ -42,7 +40,5 @@ void hstx_di_queue_tick(void);
  * @return Pointer to 36-word HSTX data island, or NULL if no packet is due.
  */
 const uint32_t *hstx_di_queue_get_audio_packet(void);
-#ifdef __cplusplus
-}
-#endif
+
 #endif // HSTX_DATA_ISLAND_QUEUE_H
