+++ drivers/pico_hdmi/video_output.c
@@ -16,9 +16,8 @@
 
 #include <math.h>
 #include <string.h>
-volatile bool HSTX_vblank = false;
-// Calculate HSTX output bit from GPIO number (GPIO12-19 => bit 0-7)
-#define HSTX_BIT_FROM_GPIO(gpio) ((gpio) - 12)
+#include "hstx.h"
+
 // ============================================================================
 // DVI/HSTX Constants
 // ============================================================================
@@ -257,23 +256,18 @@ static inline void __not_in_flash_func(video_output_handle_vsync)(dma_channel_hw
         ch->read_addr = (uintptr_t)vblank_line_vsync_on;
         ch->transfer_count = count_of(vblank_line_vsync_on);
         if (v_scanline == MODE_V_FRONT_PORCH) {
-            HSTX_vblank = true;
             video_frame_count++;
             if (vsync_callback)
                 vsync_callback();
-        } else {
-            HSTX_vblank = false;
-        }
+        } 
     } else {
         if (v_scanline == MODE_V_FRONT_PORCH) {
             ch->read_addr = (uintptr_t)vblank_acr_vsync_on;
             ch->transfer_count = vblank_acr_vsync_on_len;
             video_frame_count++;
-            HSTX_vblank = true;
             if (vsync_callback)
                 vsync_callback();
         } else {
-            HSTX_vblank = false;
             ch->read_addr = (uintptr_t)vblank_infoframe_vsync_on;
             ch->transfer_count = vblank_infoframe_vsync_on_len;
         }
@@ -623,133 +617,6 @@ void pico_hdmi_set_audio_sample_rate(uint32_t sample_rate)
     configure_audio_packets(sample_rate);
 }
 
-uint32_t pico_hdmi_getframecounter(void)
-{
-    return video_frame_count;
-}
 
-void pico_hdmi_waitForVSync(void)
-{
-#if 1
-    while (HSTX_vblank)
-    {
-        tight_loop_contents();
-    }
-    while (!HSTX_vblank)
-    {
-        tight_loop_contents();
-    }
-#endif
-}
 
-// Custom changes
-#include "pico/multicore.h" 
-#include "stdio.h"
-static uint8_t FRAMEBUFFER[(MODE_H_ACTIVE_PIXELS / 2) * (MODE_V_ACTIVE_LINES / 2) * 2];
-// uint16_t ALIGNED HDMIlines[2][MODE_H_ACTIVE_PIXELS] = {0};
-static uint8_t *WriteBuf = FRAMEBUFFER;
-static uint8_t *DisplayBuf = FRAMEBUFFER;
-static uint8_t *LayerBuf = FRAMEBUFFER;
-static uint16_t *tilefcols;
-static uint16_t *tilebcols;
-static volatile int enableScanLines = 0; // Enable scanlines
-static volatile int scanlineMode = 0;
-#define HRes (MODE_H_ACTIVE_PIXELS / 2) // 320
-#define VRes (MODE_V_ACTIVE_LINES / 2)  // 240
-uint8_t *pico_hdmi_getframebuffer(void)
-{
-    // Return the pointer to the framebuffer
-    return WriteBuf;
-}
 
-void pico_hdmi_setScanLines(int enable)
-{
-    // Set the scanlines effect flag
-    enableScanLines = enable ? 1 : 0;
-}
-
-uint16_t *__not_in_flash_func(pico_hdmi_getlineFromFramebuffer)(int scanline)
-{
-    return (uint16_t *)(WriteBuf + (scanline * HRes * 2));
-}
-void __not_in_flash_func(scanline_callbackfunc)(uint32_t v_scanline, uint32_t active_line, uint32_t *buff)
-{
-    uint16_t *p = (uint16_t *)buff;
-    int last_line = 2, load_line, line_to_load, Line_dup;
-    load_line = v_scanline - (MODE_V_TOTAL_LINES - MODE_V_ACTIVE_LINES);
-    Line_dup = load_line >> 1;
-    if (load_line >= 0 && load_line < MODE_V_ACTIVE_LINES)
-    {
-        __dmb();
-
-        for (int i = 0; i < MODE_H_ACTIVE_PIXELS / 2; i++)
-        {
-            uint8_t *d = &DisplayBuf[(Line_dup)*MODE_H_ACTIVE_PIXELS + i * 2];
-            int c = *d++;
-            c |= ((*d++) << 8);
-#if 1
-            // Apply CRT scanline effect for RGB555: darken every other line
-            if (load_line & 1 && enableScanLines)
-            { // Odd lines = scanlines
-                int r = (c >> 10) & 0x1F;
-                int g = (c >> 5) & 0x1F;
-                int b = c & 0x1F;
-                r = r >> 1;
-                g = g >> 1;
-                b = b >> 1;
-                c = (r << 10) | (g << 5) | b;
-            }
-            // tried vertical and pixel-grid scanlines
-            // but this is probably too much for the RP2350:
-#else
-            // Horizontal scanlines: darken every other line
-            if (scanlineMode == 1 && (load_line & 1))
-            {
-                int r = (c >> 10) & 0x1F;
-                int g = (c >> 5) & 0x1F;
-                int b = c & 0x1F;
-                r = r >> 1;
-                g = g >> 1;
-                b = b >> 1;
-                c = (r << 10) | (g << 5) | b;
-            }
-            // Vertical scanlines: darken every other pixel column
-            else if (scanlineMode == 2 && (i & 1))
-            {
-                int r = (c >> 10) & 0x1F;
-                int g = (c >> 5) & 0x1F;
-                int b = c & 0x1F;
-                r = r >> 1;
-                g = g >> 1;
-                b = b >> 1;
-                c = (r << 10) | (g << 5) | b;
-            }
-            // Pixel-grid: darken every other line and every other pixel column
-            else if (scanlineMode == 3 && ((load_line & 1) || (i & 1)))
-            {
-                int r = (c >> 10) & 0x1F;
-                int g = (c >> 5) & 0x1F;
-                int b = c & 0x1F;
-                r = r >> 1;
-                g = g >> 1;
-                b = b >> 1;
-                c = (r << 10) | (g << 5) | b;
-            }
-
-#endif
-            *p++ = c;
-            *p++ = c;
-        }
-    }
-}
-
-void pico_hdmi_init(void)
-{
-    hstx_di_queue_init();
-    //video_output_set_dvi_mode(true);
-    video_output_init(640, 480);
-
-    video_output_set_scanline_callback(scanline_callbackfunc);
-    multicore_launch_core1(video_output_core1_run);
-    printf("Pico HDMI initialized.\n");
-}
